refactor fsm into its own module
